+++ src/main/java/com/crud/crud/repositorie/DepartamentoRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.crud.crud.entity.Departamento;
 
 @Repository
-public interface DepartamentoRepositorio extends JpaRepository<Departamento, Integer> {
+public interface DepartamentoRepositorio extends JpaRepository<Departamento, Long> {
     
 }
